weave: force I-frames to be top
If a given frame in a makito hvec encoded stream is an I-frame then it
is guaranteed to be the /top/ line in an interlaced stream. Use this
fact to ensure that during playback, even if an odd number of frames is
lost, the stream will not remain out of sync and deinterlacing can
proceed.